requirements.txt
fastapi
uvicorn
onnxruntime==1.10.0
opencv-python==4.9.0.80
pillow==8.4.0
tqdm
python-multipart
